Refactor planner skill config creation into a separate method
Summary: This commit moves the logic of creating a planner skill config from the context variables into a separate extension method on the SKContext class. This reduces duplication and improves readability of the CreatePlanAsync method. The commit also fixes a typo in a parameter comment and adds a TODO comment for a possible improvement.
